Fix error, empty and loading states
